% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afRemover.R
\name{afMeasure}
\alias{afMeasure}
\title{Calculate textural features for each region of interest (ROI). Measures size, pixel correlation, standard deviation, and kurtosis of ROIs across two channels, output as a dataframe.}
\usage{
afMeasure(im1, im2, mask, imBit = 16)
}
\arguments{
\item{im1}{A matrix containing the intensities of image 1.}

\item{im2}{A matrix containing the intensities of image 2.}

\item{mask}{A matrix of the mask with labelled ROIs.}

\item{imBit}{bit depth of im1 and im2.}
}
\value{
A dataframe of relevant measurements for clustering.
}
\description{
Calculate textural features for each region of interest (ROI). Measures size, pixel correlation, standard deviation, and kurtosis of ROIs across two channels, output as a dataframe.
}
