% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afRemover.R
\name{intMask}
\alias{intMask}
\title{Calculate a union of the intersection masks for two masks.}
\usage{
intMask(mask1, mask2, minSize = 5)
}
\arguments{
\item{mask1}{A matrix containing a mask with annotated ROIs.}

\item{mask2}{A matrix containing a mask with annotated ROIs.}

\item{minSize}{A numeric specifying the minimum number of pixels for a ROI to be retained.}
}
\value{
A matrix of a new mask.
}
\description{
Calculate a union of the intersection masks for two masks.
}
